Clean up documentation
Cleans up documentation for clarity. Adjusted spelling error on a helper method, and adjusted all method documentation for readability and utility.
Also proposes new class, APVelocity. APVelocity is intended to be a drop-in for getting velocities as a Translation2d() object, such that users do not have to create a conversion method themselves when passing velocities into Autopilot.