Replace field-relative Translation2d with ChassisSpeeds for velocity
This lets the user just give the robot relative speeds, and because they
also give the pose each call to calculate(), we can compute the
field-relative velocity ourselves. This makes it easier on the end user.

This also makes it easier for the C++ version to secretely turn m/s to
just meters because WPILib doesn't allow arbitrary double values inside
their Translation2d types... yet (a weird chice IMO, so it's probably
C++).